// src/modules/trial-balance/components/Types/types.tsx
export type MappedRow = {
  glName: string;
  glAccount: string;
  accountType: string;
  'Level 1 Desc': string;
  'Level 2 Desc': string;
  functionalArea: string;
  amountCurrent: number;
  amountPrevious: number;
};
export type FinancialRow = {
  key: string;
  amountCurrent: number;
  amountPrevious: number;
};
export type TextRow = {
  key: string;
  amountCurrent: string;
};